sync/commits

+++ todo-backend/todo-backend.WEB/Controllers/BoardController.cs
@@ -60,10 +60,13 @@ namespace todo_backend.WEB.Controllers
             {
                 Catalog catalog = await _catalogService.GetById(el);
 
+
                 foreach (var cardId in catalog.CardsId)
                 {
                     Card card = await _cardService.GetById(cardId);
-                    await _cardService.Delete(card);
+
+                    if (card != null)
+                        await _cardService.Delete(card);
                 }
 
                 if (catalog != null)
